Moving xp and stat files
